requirements.txt
aiofiles
aiohttp
loguru
pydantic
PyYAML
transliterate
